mv qmule only package
