+++ qio/src/org/qcmg/qmule/chrconv/ChromosomeConversionRecord.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.chrconv;
+package org.qcmg.qmule.chrconv;
 
 public class ChromosomeConversionRecord {
