+++ qio/src/org/qcmg/qmule/dbsnp/Dbsnp130Record.java
@@ -9,7 +9,7 @@
 //
 
 
-package org.qcmg.dbsnp;
+package org.qcmg.qmule.dbsnp;
