+++ qio/src/org/qcmg/qmule/germlinedb/GermlineDBRecord.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.germlinedb;
+package org.qcmg.qmule.germlinedb;
 
 
 public class GermlineDBRecord {
