+++ qio/src/org/qcmg/qmule/germlinedb/GermlineDBRecordIterator.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.chrconv;
+package org.qcmg.qmule.germlinedb;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
@@ -9,48 +9,41 @@ import java.io.InputStreamReader;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 
-public class ChrConvRecordIterator implements Iterator<ChromosomeConversionRecord>{
+public final class GermlineDBRecordIterator implements Iterator<GermlineDBRecord> {
+    private final BufferedReader reader;
+    private GermlineDBRecord next;
 
-	private final BufferedReader reader;
-    private ChromosomeConversionRecord next;
+    public GermlineDBRecordIterator(final InputStream stream) {
+        InputStreamReader streamReader = new InputStreamReader(stream);
+        reader = new BufferedReader(streamReader);
+        readNext();
+    }
 
-    public ChrConvRecordIterator(final InputStream stream) {
-	        InputStreamReader streamReader = new InputStreamReader(stream);
-	        reader = new BufferedReader(streamReader);
-        	readNext();
-	    }
-
-    @Override
     public boolean hasNext() {
         return null != next;
     }
 
-    @Override
-    public ChromosomeConversionRecord next() {
+    public GermlineDBRecord next() {
         if (!hasNext()) {
             throw new NoSuchElementException();
         }
-        ChromosomeConversionRecord result = next;
+        GermlineDBRecord result = next;
         readNext();
         return result;
     }
 
     private void readNext() {
         try {
-            next = ChrConvSerializer.nextRecord(reader);
+            next = GermlineDBSerializer.nextRecord(reader);
         } catch (NoSuchElementException e) {
-        	e.printStackTrace();
             throw e;
         } catch (Exception ex) {
-        	ex.printStackTrace();
-        	next = null;
-        	throw new NoSuchElementException(ex.getMessage());
+            throw new RuntimeException(ex.getMessage(), ex);
         }
     }
 
-    @Override
+    @SuppressWarnings("unchecked")
     public void remove() {
         throw new UnsupportedOperationException();
     }
-
 }
