+++ qio/src/org/qcmg/qmule/pileup/VerifiedSnpFileReader.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.pileup;
+package org.qcmg.qmule.pileup;
 
 import java.io.Closeable;
 import java.io.File;
